move common python code into a separate directory
