+++ kiwi/__init__.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python
+## -*- python -*-
+
+#from .client import KiwiSDRStream
+#from .client import KiwiError,KiwiTooBusyError,KiwiDownError,KiwiBadPasswordError,KiwiTimeLimitError,KiwiServerTerminatedConnection,KiwiUnknownModulation
+from .client import *
+from .worker import KiwiWorker
+from .wavreader import *
